pyproject.toml
[tool.poetry]
name = "film-watch-site-api"
version = "0.1.0"
description = ""
authors = ["AbdurashiDIDO <abdurashiddido@gmail.com>"]
readme = "README.md"
packages = [{include = "film_watch_site_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-autoslug = "^1.9.9"
psycopg2-binary = "^2.9.6"
drf-yasg = "^1.21.6"
coreapi = "^2.3.3"
pyyaml = "^6.0"
celery = "^5.3.0"
pytils = "^0.4.1"
djangorestframework-simplejwt = "^5.2.2"
django-elasticsearch-dsl-drf = "^0.22.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
